feat: extract theme settings into a dedicated Theme page
Move the theme picker, design-token customization, and custom CSS out
of site Settings into a new /:site_slug/theme page (AdminLive.SiteTheme)
with its own sidebar entry, lightening Settings to identity, homepage,
custom domain, and the danger zone.

Both pages edit the same site row via Sites.update_settings; cast
ignores absent fields and validate_required falls back to existing
values, so each form safely saves only its own subset. The theme page
keeps Loader.invalidate on save.

Tests: move the theme/token suite to site_theme_live_test.exs, trim the
settings test, and assert Settings no longer renders theme controls.
